updating dependencies and trying to get to run
