--- mvvmDatabinding/src/main/java/com/bigmikehoncho/mvvmdatabinding/MvvmActivity.kt
@@ -1,9 +1,0 @@
-package com.bigmikehoncho.mvvmdatabinding
-
-import android.support.v7.app.AppCompatActivity
-
-/**
- * Created by mike on 10/4/17.
- */
-
-class MvvmActivity : AppCompatActivity()
