+++ sample/src/main/java/com/bigmikehoncho/sample/view/common/BaseActivity.kt
@@ -5,10 +5,11 @@ import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import com.bigmikehoncho.mvvmdatabinding.MvvmBinder
 import com.bigmikehoncho.mvvmdatabinding.VMBinder
+import com.bigmikehoncho.sample.view.bindings.BindingAdapters
 
 abstract class BaseActivity<Binding : ViewDataBinding> : AppCompatActivity(), VMBinder {
 
-    private val mvvmBinder = MvvmBinder<Binding>()
+    private val mvvmBinder = MvvmBinder<Binding>(BindingAdapters.defaultBinder)
     protected lateinit var binding: Binding
 
     override fun onCreate(savedInstanceState: Bundle?) {
