+++ sample/src/main/java/com/bigmikehoncho/sample/view/common/ViewModelFactory.kt
@@ -7,8 +7,8 @@ import android.arch.lifecycle.ViewModelProvider
  * Created by mike on 10/15/17.
  */
 class ViewModelFactory : ViewModelProvider.Factory {
-    
-    override fun <T : ViewModel?> create(modelClass: Class<T>?): T {
+
+    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 }
